+++ src/SumatraPDF.cpp
@@ -5161,15 +5161,18 @@ void UpdateDocumentColors() {
     static int s_lastDocumentColorsFollowTheme = -1;
     bool preservePdfImages = pagesDark && GetPreservePdfImagesInDarkMode();
     int documentColorsFollowTheme = (int)GetDocumentColorsFollowTheme();
+    bool grayscale = gSettings->fixedPageUI.grayscale;
 
     if ((text == gRenderCache->textColor) && (bg == gRenderCache->backgroundColor) &&
         (link == gRenderCache->linkColor) && preservePdfImages == s_lastPreservePdfImages &&
-        documentColorsFollowTheme == s_lastDocumentColorsFollowTheme) {
+        documentColorsFollowTheme == s_lastDocumentColorsFollowTheme &&
+        grayscale == AtomicBoolGet(&gRenderCache->grayscalePageColors)) {
         return; // colors didn't change
     }
     s_lastPreservePdfImages = preservePdfImages;
     s_lastDocumentColorsFollowTheme = documentColorsFollowTheme;
 
+    AtomicBoolSet(&gRenderCache->grayscalePageColors, grayscale);
     gRenderCache->textColor = text;
     gRenderCache->backgroundColor = bg;
     gRenderCache->linkColor = link;
@@ -13157,6 +13160,12 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
             UpdateDocumentColors();
             break;
         }
+        case CmdToggleGrayscale: {
+            gSettings->fixedPageUI.grayscale = !gSettings->fixedPageUI.grayscale;
+            UpdateDocumentColors();
+            ScheduleSaveSettings();
+            break;
+        }
 
         case CmdToggleEngineeringDrawingEnhance: {
             DisplayModel* fixedDm = win->AsFixed();
@@ -18007,6 +18016,10 @@ int APIENTRY WinMain(_In_ HINSTANCE /*hInstance*/, _In_opt_ HINSTANCE /*hPrevIns
 
     LoadSettings();
     UpdateSettings(flags);
+
+    // UpdateDocumentColors() keeps it in sync after this
+    AtomicBoolSet(&gRenderCache->grayscalePageColors, gSettings->fixedPageUI.grayscale);
+
     if (gMyWindowWasEmbedded) {
         str::ReplaceWithCopy(&gSettings->scrollbars, StrL("windows"));
     }